API 추가

+++ src/main/java/com/dreamsecurity/shopface/work/timetable/TimetableService.java
@@ -6,7 +6,7 @@ import java.util.List;
 
 public interface TimetableService {
     public boolean addTimetable(Timetable timetable, Schedule schedule);
-    public List<TimetableSchedule> getTimetableList(int branchNo);
+    public List<TimetableSchedule> getTimetableList(long branchNo);
     public boolean editTimetable(Timetable timetable, Schedule schedule);
     public boolean removeTimetable(Schedule schedule);
 }
